requirements.txt
flask==1.0.2
mysql-connector-python